Update LayerTerrain to contain a finalMap
The finalMap variable will hold the values of all of the combined layers and work as the base for the Deform class.
Deform class is also stubbed out with a DeformTerrain function that needs something to signal whichever terrain class is generating the terrain.
Consider making the two generators inherit from a base class at some point if they continue to diverge.