Terrain is Deforming!
Holy hell there were so many variables for height and width that were being assigned incorrectly in the terrain generation. I very much need to go through it and standardize the order they are used in.
The terrain can now have the deform
script run on it but it does redraw the entire mesh and it needs to only update the chunk that is operated on.
The craters are squares due to the manhattan distance heuristic used, so next steps will be adding circular distances and some other shapes.