Textures (#20)
* this is borked, pushing for vic to fix

* IT'S WORKING!!! proper splatmaps, saving before I start integrating the moisture map for coloring the splatmap.

* Correctly lerping to get 0 -1 values. Saving here before working on variable splatmaps :)

* dirtty solutions, but solutions nonetheless. Power amd minvalues are working. Something is bugged with the layering, adding layers changes the amount of land and adds detail.

* implemented moisture map into biome coloring.

* Added Generate New Seed button.

* fucky wucky

reeee

* fixing commits

* added gen new seed button, scene updated woot woot

let's goooo

* Textures are updated on terrain change

Refactored texture process. Now we can give it an explicit X,Y to change textures.

* fixed splatmap bug

fixed bug where map was not being updated on a deform when generating new map seed. reeeeee

* camera rotation on both axis'

* uhhhh

* Adding some models for trees/rocks. Texture files accessible by name of file as key in dict.

Reworking how we are accessing textures, so that we can access them by name in a dictionary in the code instead of some arbitrary index in an array.
This can also be applied to biome models, like trees and rocks to make things easier.

* Textures loading and accessing when setting textures is working. Still need to plug into Vicky's biome values.